feat: use workload context for metric data
By using the workload context in the metrics it is easier to create dashboards per workload. We also ran into issues when using this solution in larger setups. So we are now using stepfunction for pagination to avoid limits. By using stepfunctions to iterate over the pages, we remove the need of a sleep in the lambda. (To avoid throttling on the Security Hub API) Because we can now rely on the retry logic of stepfunctions. Also, the lambda invocation timeout is removed due to this change (15 minutes) as we only invoke the function for each page.