feat: collect all existing accounts
It might occur that an account does not have any violation in SecurityHub. When we run a report for a conformance pack it means that the account is 100% compliance. But because there is no finding we don't know about the account. We can solve this by fetching all accounts from the AWS Organization and use that as a basis.