refactor: query only the failed findings
By querying only the failed findings we reduce the amount of findings that we need to fetch. This would lower the chances of running into service limits. And it decreases the costs because we need to move less data.
Because we now only have the failed items we need to know the total amount of enabled controls for this standard.